Cambiadas claves primarias de Int a Long para que al insertar se pueda devolvel la clave autogenerada. Relacion 1N entre medicamento y medicamentoActivo. Pendiente guardar las imagenes en local en vez de en BBDD

+++ ProxectoFCT/app/src/main/java/com/a23pablooc/proxectofct/data/network/CimaApiClient.kt
@@ -11,7 +11,7 @@ import retrofit2.http.Query
 interface CimaApiClient {
     @GET("rest/medicamento")
     suspend fun getMedicamentoByCodNacional(
-        @Query("cn") cn: Int
+        @Query("cn") cn: Long
     ): Response<MedicamentoModel>
 
     @GET("rest/medicamento")
