+++ ProxectoFCT/app/src/main/java/com/a23pablooc/proxectofct/domain/model/MedicamentoActivoItem.kt
@@ -3,7 +3,7 @@ package com.a23pablooc.proxectofct.domain.model
 import java.util.Date
 
 data class MedicamentoActivoItem(
-    var pkMedicamentoActivo: Int,
+    var pkMedicamentoActivo: Long,
     var fkMedicamento: MedicamentoItem,
     var fechaInicio: Date,
     var fechaFin: Date,
