+++ ProxectoFCT/app/src/main/java/com/a23pablooc/proxectofct/ui/view/viewholders/AddActiveMedDialogViewModel.kt
@@ -17,6 +17,6 @@ class AddActiveMedDialogViewModel @Inject constructor(
         if (!codNacionalPattern.matches(codNacional))
             throw IllegalArgumentException("Invalid codNacional")
 
-        return searchMedicamentoUseCase.invoke(codNacional.substringBefore('.').toInt())
+        return searchMedicamentoUseCase.invoke(codNacional.substringBefore('.').toLong())
     }
 }
